Revert "changed amount to int"
This reverts commit 6ec4403356dc7c1e16847a3ba978ae441db61646.

+++ src/main/kotlin/com/itmo/microservices/demo/orders/api/service/OrderService.kt
@@ -10,7 +10,7 @@ import java.util.*
 
 interface OrderService {
     fun createOrder() : OrderDto
-    fun putItemToOrder(orderId : UUID, itemId : UUID, amount : Int) : ResponseEntity<Nothing>
+    fun putItemToOrder(orderId : UUID, itemId : UUID, amount : Long) : ResponseEntity<Nothing>
     fun getOrder(orderId: UUID) :OrderDto
 //    fun getOrdersByUsername(user : UserDetails) : List<OrderModel>
 //    fun getOrder(orderId : UUID) : OrderModel
