+++ src/main/kotlin/com/itmo/microservices/demo/orders/impl/util/OrderEntityModelMappers.kt
@@ -2,7 +2,9 @@ package com.itmo.microservices.demo.orders.impl.util
 
 import com.itmo.microservices.demo.orders.api.model.BookingDto
 import com.itmo.microservices.demo.orders.api.model.OrderDto
+import com.itmo.microservices.demo.orders.api.model.OrderModel
 import com.itmo.microservices.demo.orders.impl.entity.Order
+import com.itmo.microservices.demo.stock.impl.entity.StockItem
 import java.util.*
 
 //fun OrderModel.toEntity() = Order(
@@ -20,7 +22,7 @@ import java.util.*
 //    userId = this.userId,
 //    status = this.status
 //)
-fun Order.toDto(itemsMap: Map<UUID, Int>) : OrderDto = OrderDto(
+fun Order.toDto(itemsMap: Map<UUID, Long>) : OrderDto = OrderDto(
     id = this.id,
     timeCreated = this.timeCreated,
     status = this.status,
